Fixed conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@66795 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/data_loaders/genbank/reader_snp.cpp
@@ -181,7 +181,7 @@ CSNP_Seq_feat_hook::~CSNP_Seq_feat_hook(void)
 {
     if ( CollectSNPStat() ) {
         size_t total =
-            accumulate(m_Count, m_Count+SSNP_Info::eSNP_Type_last, 0);
+            accumulate(m_Count, m_Count+SSNP_Info::eSNP_Type_last, size_t(0));
         NcbiCout << "CSeq_annot_SNP_Info statistic (gi = " <<
             m_Seq_annot_SNP_Info.GetGi() << "):\n";
         for ( size_t i = 0; i < SSNP_Info::eSNP_Type_last; ++i ) {
@@ -194,7 +194,7 @@ CSNP_Seq_feat_hook::~CSNP_Seq_feat_hook(void)
         NcbiCout << NcbiEndl;
 
         total =
-            accumulate(s_TotalCount, s_TotalCount+SSNP_Info::eSNP_Type_last,0);
+            accumulate(s_TotalCount, s_TotalCount+SSNP_Info::eSNP_Type_last,size_t(0));
         NcbiCout << "cumulative CSeq_annot_SNP_Info statistic:\n";
         for ( size_t i = 0; i < SSNP_Info::eSNP_Type_last; ++i ) {
             NcbiCout <<
